+++ test/geometry_verificator/geometry_verificator.cpp
@@ -85,7 +85,7 @@ int main(int argc, char **argv)
     /*
      * Y: Output distances to (0, 0, 0)
      */
-    for (int i = 0; i < job.photostation.cameras.size(); ++i)
+    for (size_t i = 0; i < job.photostation.cameras.size(); ++i)
     {
         auto cam = job.photostation.getRawCamera(map_fwd[i]);
         std::cout << "|_." << cam.nameId << "|" << !cam.extrinsics.position << "|" << std::endl;
@@ -163,7 +163,7 @@ int main(int argc, char **argv)
         auto cam = job.photostation.getRawCamera(map_fwd[i]);
         raysHigh.emplace_back(cam.rayFromPixel(cam.intrinsics.principal));
     }
-    for (int i = 7; i < job.photostation.cameras.size(); ++i)
+    for (size_t i = 7; i < job.photostation.cameras.size(); ++i)
     {
         auto cam = job.photostation.getRawCamera(map_fwd[i]);
         raysLow.emplace_back(cam.rayFromPixel(cam.intrinsics.principal));
